refactor(replsession): split rewrite.go into rewrite + static_report, remove dead PersistPolicy.Sessions (CQ-2, CQ-6)
CQ-2: Move static analysis report building functions from rewrite.go
into static_report.go. The rewrite file now contains only the IIFE
source rewriting logic; the new file handles AST/CST/scope report
construction.

Functions moved: buildStaticReport, rangeFromNode, declarationSnippet,
bindingReferences, buildScopeView, expandAllNodes, inspectorRows,
buildRows, flattenCST, trimForDisplay, rowLike type.

CQ-6: Remove PersistPolicy.Sessions field which was set to true in
PersistentSessionOptions() but never read by any code path.
Note: PersistPolicy.Evaluations IS read in persistence.go:21, so
the original review was partially incorrect — only Sessions is dead.

GOJA-044 CQ-2 CQ-6
